feat: use common router
