+++ src/components/common/share-modal.tsx
@@ -114,51 +114,51 @@ const Share = ({
             </li>
 
             <ul>
-              {recipients.length > 0 && (
-                <Fragment>
-                  <li className="group flex items-center justify-between gap-x-2 py-2">
-                    <div className="flex items-center gap-2">
-                      <Avatar className="h-8 w-8 rounded-full">
-                        <AvatarImage src={"/placeholders/user.svg"} />
-                      </Avatar>
-                      <span className="flex flex-col font-medium">
-                        <span className="text-sm">John Doe</span>
-                        <span className="text-xs text-primary/50">Member</span>
-                      </span>
-                    </div>
+              {/* {recipients.length > 0 && ( */}
+              <Fragment>
+                <li className="group flex items-center justify-between gap-x-2 py-2">
+                  <div className="flex items-center gap-2">
+                    <Avatar className="h-8 w-8 rounded-full">
+                      <AvatarImage src={"/placeholders/user.svg"} />
+                    </Avatar>
+                    <span className="flex flex-col font-medium">
+                      <span className="text-sm">John Doe</span>
+                      <span className="text-xs text-primary/50">Member</span>
+                    </span>
+                  </div>
 
-                    <div className="ml-auto content-end justify-end object-right">
-                      <div className="flex items-center gap-1">
-                        <Button
-                          size="sm"
-                          variant={"secondary"}
-                          className=" hidden text-sm font-medium text-red-500/70 hover:text-red-500 group-hover:flex"
-                        >
-                          <DeleteIcon className="h-4 w-4" />
-                          Remove
-                        </Button>
-                        <Button
-                          size="sm"
-                          variant={"secondary"}
-                          className="flex text-sm font-medium text-primary/70 hover:text-primary/90"
-                          onClick={() =>
-                            copyToClipboard(2, "https://example.com/1")
-                          }
-                        >
-                          <Fragment>
-                            {copiedIdx === 2 ? (
-                              <CheckIcon className="h-4 w-4 text-green-500" />
-                            ) : (
-                              <LinkIcon className="h-4 w-4" />
-                            )}
-                            Copy link
-                          </Fragment>
-                        </Button>
-                      </div>
+                  <div className="ml-auto content-end justify-end object-right">
+                    <div className="flex items-center gap-1">
+                      <Button
+                        size="sm"
+                        variant={"secondary"}
+                        className=" hidden text-sm font-medium text-red-500/70 hover:text-red-500 group-hover:flex"
+                      >
+                        <DeleteIcon className="h-4 w-4" />
+                        Remove
+                      </Button>
+                      <Button
+                        size="sm"
+                        variant={"secondary"}
+                        className="flex text-sm font-medium text-primary/70 hover:text-primary/90"
+                        onClick={() =>
+                          copyToClipboard(2, "https://example.com/1")
+                        }
+                      >
+                        <Fragment>
+                          {copiedIdx === 2 ? (
+                            <CheckIcon className="h-4 w-4 text-green-500" />
+                          ) : (
+                            <LinkIcon className="h-4 w-4" />
+                          )}
+                          Copy link
+                        </Fragment>
+                      </Button>
                     </div>
-                  </li>
-                </Fragment>
-              )}
+                  </div>
+                </li>
+              </Fragment>
+              {/* )} */}
             </ul>
 
             <div className="mt-6 flex justify-end">
