+++ src/types/contacts.ts
@@ -13,4 +13,5 @@ export type ShareRecipient = {
   id: string;
   name: string;
   email: string;
+  image?: string;
 };
